Splitt server.js i lokal og produksjon for å kunne kjøre opp i prod uten webpack-config
